Extract OWL browser core library
