+++ demos/owl-render-fixture/CHROMIUM.md
@@ -31,10 +31,12 @@ then Swift hosts the portal id in `CALayerHost`.
 `OwlFreshSession`, and the child `OwlFreshProfile`, `OwlFreshWebView`,
 `OwlFreshInput`, `OwlFreshSurfaceTreeHost`, and `OwlFreshNativeSurfaceHost`
 surfaces. `OwlMojoBindingsGenerator` emits
-`Sources/OwlMojoBindingsGenerated/OwlFresh.generated.swift`, and the verifier
-uses those generated Swift request/event, transport, pending-handle bind graph,
-and `MojoPipeBindings` types before calling typed runtime symbols. The old
-generic `interface + method + JSON` invoke bridge is not used.
+`Sources/OwlMojoBindingsGenerated/OwlFresh.generated.swift`. `OwlBrowserCore`
+owns the Swift runtime loader, session events, generated pending-handle bind
+graph, and typed browser commands before calling typed runtime symbols. The
+verifier imports `OwlBrowserCore` and owns only AppKit hosting, screenshots,
+fixtures, and artifact reporting. The old generic `interface + method + JSON`
+invoke bridge is not used.
 
 The verified gate now includes input. `run-layer-host-verifier-gui.sh` can run
 the real Chromium compositor input fixtures with `OWL_LAYER_HOST_INPUT_CHECK=1`;
