+++ demos/owl-render-fixture/README.md
@@ -2,11 +2,11 @@
 
 This verifier is the first visual gate for the OWL effort. It launches a Chromium host executable on macOS, renders deterministic HTML fixtures, writes PNG artifacts, and checks pixel colors in the captured images.
 
-`OwlLayerHostVerifier` is the current first architecture gate. Swift loads the OWL Mojo runtime dylib, launches Content Shell through Mojo, waits for Chromium to publish a browser-process portal `CAContext` id over Mojo, hosts that id in a native `CALayerHost`, screenshots the Swift-owned host window, and checks the pixels. Swift host requests go through generated transport and pipe binding types from `Mojo/OwlFresh.mojom`: `OwlFreshSession` owns lifetime and client binding, while `OwlFreshProfile`, `OwlFreshWebView`, `OwlFreshInput`, `OwlFreshSurfaceTreeHost`, and `OwlFreshNativeSurfaceHost` own their respective browser surfaces. The verifier no longer uses the old generic `interface + method + JSON` invoke bridge. Generated `MojoPipeBindings` now forward typed pending-handle bind calls for `SetClient` and `Bind*`, then forward typed runtime calls such as resize, key input, surface capture, and popup acceptance. `Scripts/run-layer-host-fixture-verifier-gui.sh` uses a Chromium-owned layer fixture context to prove the Mojo plus native display plumbing. `Scripts/run-layer-host-verifier-gui.sh` uses the real Chromium compositor portal and runs a deterministic canvas fixture plus `https://example.com/` by default. Set `OWL_LAYER_HOST_INPUT_CHECK=1` to also verify real mouse/key input changes web content from `OWL_INPUT_READY` to `OWL_INPUT_CLICKED`, type into a form field, check a checkbox, click submit, verify Command, Option, Control, and Shift modifier key delivery, scroll to a specific numbered content row with a Mojo wheel event, exercise text editing plus selection replacement semantics, and publish native popup surfaces through the Mojo surface tree. Set `OWL_LAYER_HOST_RESIZE_CHECK=1` to verify resize-small and resize-roundtrip by resizing both Chromium web contents and the Swift `CALayerHost` window.
+`OwlLayerHostVerifier` is the current first architecture gate. Swift loads the OWL Mojo runtime dylib through the `OwlBrowserCore` library, launches Content Shell through Mojo, waits for Chromium to publish a browser-process portal `CAContext` id over Mojo, hosts that id in a native `CALayerHost`, screenshots the Swift-owned host window, and checks the pixels. Swift host requests go through generated transport and pipe binding types from `Mojo/OwlFresh.mojom`: `OwlFreshSession` owns lifetime and client binding, while `OwlFreshProfile`, `OwlFreshWebView`, `OwlFreshInput`, `OwlFreshSurfaceTreeHost`, and `OwlFreshNativeSurfaceHost` own their respective browser surfaces. The verifier no longer uses the old generic `interface + method + JSON` invoke bridge. Generated `MojoPipeBindings` now forward typed pending-handle bind calls for `SetClient` and `Bind*`, then forward typed runtime calls such as resize, key input, surface capture, and popup acceptance. `Scripts/run-layer-host-fixture-verifier-gui.sh` uses a Chromium-owned layer fixture context to prove the Mojo plus native display plumbing. `Scripts/run-layer-host-verifier-gui.sh` uses the real Chromium compositor portal and runs a deterministic canvas fixture plus `https://example.com/` by default. Set `OWL_LAYER_HOST_INPUT_CHECK=1` to also verify real mouse/key input changes web content from `OWL_INPUT_READY` to `OWL_INPUT_CLICKED`, type into a form field, check a checkbox, click submit, verify Command, Option, Control, and Shift modifier key delivery, scroll to a specific numbered content row with a Mojo wheel event, exercise text editing plus selection replacement semantics, and publish native popup surfaces through the Mojo surface tree. Set `OWL_LAYER_HOST_RESIZE_CHECK=1` to verify resize-small and resize-roundtrip by resizing both Chromium web contents and the Swift `CALayerHost` window.
 
 ## Current status
 
-The real compositor path is Mojo plus Chromium-owned `CAContext` plus Swift `CALayerHost`. It does not use Unix sockets or remote debugging. The current Chromium patch adds `fresh-owl-hosted-frame-pump`, a scoped Owl switch that maps hosted surfaces into Chromium's existing renderer, root compositor, and GPU frame-pump settings without passing the broad `--disable-frame-rate-limit` command-line switch.
+The real compositor path is Mojo plus Chromium-owned `CAContext` plus Swift `CALayerHost`. It does not use Unix sockets or remote debugging. The current Swift shape has an `OwlBrowserCore` library for runtime loading, session events, generated bind graph ownership, and typed browser commands; `OwlLayerHostVerifier` is now a client that owns AppKit hosting, screenshots, fixture orchestration, and artifact reporting. The current Chromium patch adds `fresh-owl-hosted-frame-pump`, a scoped Owl switch that maps hosted surfaces into Chromium's existing renderer, root compositor, and GPU frame-pump settings without passing the broad `--disable-frame-rate-limit` command-line switch.
 
 The current visual gates are intentionally small but behavioral: example.com, deterministic canvas, click, form typing, modifier keys, resize-small, resize-roundtrip, scroll, text-edit selection replacement, widget controls, a browser-default `<select>` with no select CSS, collapsed `<select>` popup publication, right-click context menu publication, and live Google search-box typing. `Scripts/run-layer-host-focused-suites-gui.sh` is the default broad gate: it splits the checks into focused render, input, resize, scroll-text, widgets, and Google batches, writes one artifact directory per suite, and emits a screenshot checklist at `artifacts/layer-host-focused-gui-latest/focused-suites.txt`. The old full all-target run is still useful as a stress test, but it is not the default pass/fail gate because it is flaky after many sequential sessions.
 
@@ -14,10 +14,10 @@ The current visual gates are intentionally small but behavioral: example.com, de
 
 1. Replace the patch-file workflow with an owned Chromium fork or DEPS-pinned patch application step that can be built reproducibly in CI.
 2. Promote the focused-suite runner to CI on the AWS Mac so each suite reports its own summary and visual artifact bundle.
-3. Replace the remaining dylib bootstrap boundary with a production embeddable Swift client library while keeping the current generated `OwlFreshSession`, pending-handle bind graph, and child-interface shape.
+3. Replace the remaining dylib bootstrap boundary inside `OwlBrowserCore` with an app-integrated browser runtime while keeping the current generated `OwlFreshSession`, pending-handle bind graph, and child-interface shape.
 4. Add popup/native widget coverage for color pickers, permission prompts, file pickers, extension bubbles, authentication prompts, and other separate native or `RenderWidgetHostView` surfaces.
 5. Add lifecycle coverage for tab attach/detach, hidden/visible transitions, crash/restart recovery, device scale changes, and cross-display moves.
-6. Move from test harness Swift code toward an embeddable client library API with the same generated Mojo transport.
+6. Continue moving verifier-only behavior out of `OwlBrowserCore`, keeping screenshots, pixels, and fixture HTML in the verifier while the library owns reusable browser sessions.
 
 `OwlLayerHostSelfTest` is the smallest local rendering gate. Direct mode draws deterministic red, green, blue, and text layers into a normal Swift layer-backed `NSWindow`, proving the window capture environment. Layer-host mode creates a private `CAContext` in Swift, draws the same layers into it, hosts that context in the same Swift `CALayerHost` window path, screenshots the window, and checks the pixels. This isolates whether `CAContext` plus `CALayerHost` plus screenshot capture works before involving Chromium.
